Move bucket containers to their own directory.
