+++ scripts/containers/bucket/BucketCollectionsPage.js
@@ -2,9 +2,9 @@ import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import { push as updatePath } from "react-router-redux";
 
-import BucketCollections from "../components/bucket/BucketCollections";
-import * as BucketActions from "../actions/bucket";
-import * as NotificationsActions from "../actions/notifications";
+import BucketCollections from "../../components/bucket/BucketCollections";
+import * as BucketActions from "../../actions/bucket";
+import * as NotificationsActions from "../../actions/notifications";
 
 
 function mapStateToProps(state) {
