+++ scripts/containers/bucket/BucketCreatePage.js
@@ -1,15 +1,14 @@
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 
-import BucketEditForm from "../components/bucket/BucketEditForm";
-import * as BucketActions from "../actions/bucket";
+import BucketCreateForm from "../../components/bucket/BucketCreateForm";
+import * as BucketActions from "../../actions/bucket";
 
 
 function mapStateToProps(state) {
   return {
     session: state.session,
     bucket: state.bucket,
-    capabilities: state.session.serverInfo.capabilities,
   };
 }
 
@@ -20,4 +19,4 @@ function mapDispatchToProps(dispatch) {
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(BucketEditForm);
+)(BucketCreateForm);
