+++ scripts/containers/bucket/BucketEditPage.js
@@ -1,14 +1,15 @@
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 
-import BucketCreateForm from "../components/bucket/BucketCreateForm";
-import * as BucketActions from "../actions/bucket";
+import BucketEditForm from "../../components/bucket/BucketEditForm";
+import * as BucketActions from "../../actions/bucket";
 
 
 function mapStateToProps(state) {
   return {
     session: state.session,
     bucket: state.bucket,
+    capabilities: state.session.serverInfo.capabilities,
   };
 }
 
@@ -19,4 +20,4 @@ function mapDispatchToProps(dispatch) {
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(BucketCreateForm);
+)(BucketEditForm);
