+++ scripts/containers/bucket/BucketGroupsPage.js
@@ -1,9 +1,10 @@
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
-import BucketGroups from "../components/bucket/BucketGroups";
-import * as BucketActions from "../actions/bucket";
-import * as NotificationsActions from "../actions/notifications";
 import { push as updatePath } from "react-router-redux";
+
+import BucketGroups from "../../components/bucket/BucketGroups";
+import * as BucketActions from "../../actions/bucket";
+import * as NotificationsActions from "../../actions/notifications";
 
 
 function mapStateToProps(state) {
