+++ scripts/containers/bucket/BucketHistoryPage.js
@@ -2,9 +2,9 @@ import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import { push as updatePath } from "react-router-redux";
 
-import BucketHistory from "../components/bucket/BucketHistory";
-import * as BucketActions from "../actions/bucket";
-import * as NotificationsActions from "../actions/notifications";
+import BucketHistory from "../../components/bucket/BucketHistory";
+import * as BucketActions from "../../actions/bucket";
+import * as NotificationsActions from "../../actions/notifications";
 
 
 function mapStateToProps(state) {
