+++ scripts/routes.js
@@ -10,11 +10,11 @@ import App from "./containers/App";
 import HomePage from "./containers/HomePage";
 import Sidebar from "./containers/Sidebar";
 import Notifications from "./containers/Notifications";
-import BucketCreatePage from "./containers/BucketCreatePage";
-import BucketEditPage from "./containers/BucketEditPage";
-import BucketCollectionsPage from "./containers/BucketCollectionsPage";
-import BucketGroupsPage from "./containers/BucketGroupsPage";
-import BucketHistoryPage from "./containers/BucketHistoryPage";
+import BucketCreatePage from "./containers/bucket/BucketCreatePage";
+import BucketEditPage from "./containers/bucket/BucketEditPage";
+import BucketCollectionsPage from "./containers/bucket/BucketCollectionsPage";
+import BucketGroupsPage from "./containers/bucket/BucketGroupsPage";
+import BucketHistoryPage from "./containers/bucket/BucketHistoryPage";
 import GroupCreatePage from "./containers/GroupCreatePage";
 import GroupEditPage from "./containers/GroupEditPage";
 import GroupHistoryPage from "./containers/GroupHistoryPage";
